updates to analysis and packages
