+++ singularity/views/__init__.py
@@ -1,0 +1,10 @@
+from .tree import (
+    container_difference as container_difference_tree,
+    container_similarity as container_similarity_tree,
+    container_tree,
+    make_container_tree,
+    make_package_tree,
+    make_interactive_tree
+)
+
+
